requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==2.0
idna==2.6
lxml==3.8.0
ncclient==0.5.3
paramiko==2.2.1
pyang==1.7.3
pyasn1==0.3.3
pycparser==2.18
PyNaCl==1.1.2
six==1.10.0
pyyaml==3.11
